+++ test/core/db/tips/model/tip_test.dart
@@ -2,7 +2,7 @@ import 'package:der_die_das/core/db/tips/model/tip.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import '../../../test_utils.dart';
+import '../../../../test_utils.dart';
 
 void main() {
   testWidgets('getAllTips', (tester) async {
